+++ src/redturtle/volto/__init__.py
@@ -1,17 +1,15 @@
 # -*- coding: utf-8 -*-
 """Init and utils."""
-import logging
-import re
-
 from plone.app.content.browser.vocabulary import PERMISSIONS
 from plone.folder.nogopip import GopipIndex
-from plone.restapi.serializer import utils
 from Products.ZCatalog.Catalog import Catalog
+from redturtle.volto.catalogplan import Catalog_sorted_search_indexes
 from zope.i18nmessageid import MessageFactory
 from ZTUtils.Lazy import LazyCat
 from ZTUtils.Lazy import LazyMap
 
-from redturtle.volto.catalogplan import Catalog_sorted_search_indexes
+import logging
+
 
 logger = logging.getLogger(__name__)
 
@@ -77,6 +75,3 @@ def Catalog_sortResults(
 logger.info("install monkey patch for Products.ZCatalog.Catalog.Catalog.sortResults")
 Catalog._orig_sortResults = Catalog.sortResults
 Catalog.sortResults = Catalog_sortResults
-
-# patch plone.restapi regexp to catch also other
-utils.RESOLVEUID_RE = re.compile("^(?:|.*/)resolve[Uu]id/([^/]*)/?(.*)$")