+++ src/redturtle/volto/restapi/search/query.py
@@ -1,16 +1,15 @@
 # -*- coding: utf-8 -*-
-import re
-
 from plone.restapi.interfaces import IIndexQueryParser
 from plone.restapi.search.query import (
     ZCTextIndexQueryParser as BaseZCTextIndexQueryParser,
 )
 from Products.ZCTextIndex.ZCTextIndex import ZCTextIndex
-from zope.component import adapter
-from zope.interface import Interface
-from zope.interface import implementer
-
 from redturtle.volto.interfaces import IRedturtleVoltoLayer
+from zope.component import adapter
+from zope.interface import implementer
+from zope.interface import Interface
+
+import re
 
 
 @implementer(IIndexQueryParser)