+++ src/redturtle/volto/tests/test_blocks_searchable_text.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
-import unittest
-
 from plone import api
-from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from redturtle.volto.testing import REDTURTLE_VOLTO_FUNCTIONAL_TESTING
 from transaction import commit
 
-from redturtle.volto.testing import REDTURTLE_VOLTO_FUNCTIONAL_TESTING
+import unittest
 
 
 class TestBlocksSearchable(unittest.TestCase):